+++ temba/ai/views.py
@@ -14,12 +14,23 @@ class LLMCRUDL(SmartCRUDL):
     actions = ("list", "delete")
 
     class List(SpaMixin, ContextMenuMixin, BaseListView):
-        title = _("AI Models")
+        title = _("Artificial Intelligence")
         menu_path = "settings/ai"
         default_order = ("name",)
 
+        def get_context_data(self, **kwargs):
+            context = super().get_context_data(**kwargs)
+
+            count, limit = LLM.get_org_limit_progress(self.request.org)
+            context["llm_limit"] = limit
+            context["llm_count"] = count
+
+            return context
+
         def build_context_menu(self, menu):
-            if self.has_org_perm("ai.llm_connect"):
+            count, limit = LLM.get_org_limit_progress(self.request.org)
+
+            if self.has_org_perm("ai.llm_connect") and count < limit:
                 menu.add_modax(
                     _("New"),
                     "new-llm",