+++ temba/orgs/models.py
@@ -258,6 +258,7 @@ class Org(SmartModel):
     LIMIT_GLOBALS = "globals"
     LIMIT_GROUPS = "groups"
     LIMIT_LABELS = "labels"
+    LIMIT_LLMS = "llms"
     LIMIT_TOPICS = "topics"
     LIMIT_TEAMS = "teams"
 
@@ -634,9 +635,9 @@ class Org(SmartModel):
         return self.get_call_channel() or self.get_answer_channel()
 
     def is_outbox_full(self) -> bool:
-        from temba.msgs.models import SystemLabel
+        from temba.msgs.models import MsgFolder
 
-        return SystemLabel.get_counts(self)[SystemLabel.TYPE_OUTBOX] >= 1_000_000
+        return MsgFolder.OUTBOX.get_count(self) >= 1_000_000
 
     def get_estimated_send_time(self, msg_count):
         """