+++ temba/utils/views/wizard.py
@@ -18,7 +18,8 @@ class SmartWizardView(SmartView, SessionWizardView):
         return []
 
     def lookup_field_help(self, field, default=None):
-        return self.form_list[self.steps.current].base_fields[field].help_text
+        form = self.get_form(self.steps.current)
+        return form.fields[field].help_text or default
 
     def lookup_field_label(self, context, field, object):
         return context["form"].fields[field].label